refactor: move translationWithClamping to Cesium/utils
