+++ src/components/molecules/Visualizer/Engine/Cesium/utils.ts
@@ -1,4 +1,10 @@
-import { Cartesian3, Viewer as CesiumViewer, Math as CesiumMath } from "cesium";
+import {
+  Cartesian3,
+  Viewer as CesiumViewer,
+  Math as CesiumMath,
+  TranslationRotationScale,
+  Cartographic,
+} from "cesium";
 
 export const convertCartesian3ToPosition = (
   cesium?: CesiumViewer,
@@ -12,4 +18,20 @@ export const convertCartesian3ToPosition = (
     lng: CesiumMath.toDegrees(cartographic.longitude),
     height: cartographic.height,
   };
+};
+
+export const translationWithClamping = (
+  trs: TranslationRotationScale,
+  keepBoxAboveGround: boolean,
+  terrainHeightEstimate: number,
+) => {
+  if (keepBoxAboveGround) {
+    const cartographic = Cartographic.fromCartesian(trs.translation, undefined, new Cartographic());
+    const boxBottomHeight = cartographic.height - trs.scale.z / 2;
+    const floorHeight = terrainHeightEstimate;
+    if (boxBottomHeight < floorHeight) {
+      cartographic.height += floorHeight - boxBottomHeight;
+      Cartographic.toCartesian(cartographic, undefined, trs.translation);
+    }
+  }
 };
